requirements.txt
google-api-python-client>=2.100.0,<3
google-auth-oauthlib>=1.2.0,<2
google-auth-httplib2>=0.2.0,<1
python-docx>=1.1.0,<2
